pyproject.toml
[project]
name = "ocrproject"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "ollama>=0.6.1",
    "pillow>=12.2.0",
]
